weaviate_1/requirements.txt
streamlit
streamlit-option-menu
langchain
langchain_community
weaviate-client
pyspark
python-dotenv
